+++ tfhe/src/core_crypto/algorithms/lwe_multi_bit_bootstrap_key_generation.rs
@@ -708,6 +708,49 @@ pub fn par_generate_seeded_lwe_multi_bit_bootstrap_key<
     // Maybe Sized allows to pass Box<dyn Seeder>.
     NoiseSeeder: Seeder + ?Sized,
 {
+    let mut generator = EncryptionRandomGenerator::<DefaultRandomGenerator>::new(
+        output.compression_seed().seed,
+        noise_seeder,
+    );
+
+    par_generate_seeded_lwe_multi_bit_bootstrap_key_with_pre_seeded_generator(
+        input_lwe_secret_key,
+        output_glwe_secret_key,
+        output,
+        noise_distribution,
+        &mut generator,
+    );
+}
+
+/// Parallel variant of [`generate_seeded_lwe_multi_bit_bootstrap_key_with_pre_seeded_generator`],
+/// it is recommended to use this function for better key generation times as LWE bootstrapping
+/// keys can be quite large.
+///
+/// This uses an already seeded generator
+#[allow(clippy::too_many_arguments)]
+pub fn par_generate_seeded_lwe_multi_bit_bootstrap_key_with_pre_seeded_generator<
+    InputScalar,
+    OutputScalar,
+    NoiseDistribution,
+    InputKeyCont,
+    OutputKeyCont,
+    OutputCont,
+    ByteGen,
+>(
+    input_lwe_secret_key: &LweSecretKey<InputKeyCont>,
+    output_glwe_secret_key: &GlweSecretKey<OutputKeyCont>,
+    output: &mut SeededLweMultiBitBootstrapKey<OutputCont>,
+    noise_distribution: NoiseDistribution,
+    generator: &mut EncryptionRandomGenerator<ByteGen>,
+) where
+    InputScalar: UnsignedInteger + CastFrom<usize> + CastInto<OutputScalar> + Sync,
+    OutputScalar: Encryptable<Uniform, NoiseDistribution> + Sync + Send,
+    NoiseDistribution: Distribution + Sync,
+    InputKeyCont: Container<Element = InputScalar>,
+    OutputKeyCont: Container<Element = OutputScalar> + Sync,
+    OutputCont: ContainerMut<Element = OutputScalar>,
+    ByteGen: ByteRandomGenerator + ParallelByteRandomGenerator,
+{
     assert!(
         output.input_lwe_dimension() == input_lwe_secret_key.lwe_dimension(),
         "Mismatched LweDimension between input LWE secret key and LWE bootstrap key. \
@@ -730,11 +773,6 @@ pub fn par_generate_seeded_lwe_multi_bit_bootstrap_key<
         Output GLWE secret key PolynomialSize: {:?}, LWE bootstrap key PolynomialSize {:?}.",
         output_glwe_secret_key.polynomial_size(),
         output.polynomial_size()
-    );
-
-    let mut generator = EncryptionRandomGenerator::<DefaultRandomGenerator>::new(
-        output.compression_seed().seed,
-        noise_seeder,
     );
 
     let forking_configuration = output.encryption_fork_config(Uniform, noise_distribution);
